[python] Move temporary feature-flag into a module
