+++ apis/python/src/tiledbsoma/__init__.py
@@ -97,18 +97,6 @@ import ctypes
 import os
 import sys
 
-# Temporary for https://github.com/single-cell-data/TileDB-SOMA/issues/2407
-_new_shape_feature_flag = os.getenv("SOMA_PY_NEW_SHAPE") is not None
-
-
-def _new_shape_feature_flag_enabled() -> bool:
-    """
-    This is temporary only and will be removed once
-    https://github.com/single-cell-data/TileDB-SOMA/issues/2407
-    is complete.
-    """
-    return _new_shape_feature_flag
-
 
 # Load native libraries. On wheel builds, we may have a shared library
 # already linked. In this case, we can import directly
