+++ apis/python/src/tiledbsoma/_dataframe.py
@@ -14,7 +14,7 @@ import somacore
 from somacore import options
 from typing_extensions import Self
 
-from tiledbsoma import _new_shape_feature_flag_enabled
+from tiledbsoma._flags import _new_shape_feature_flag_enabled
 
 from . import _arrow_types, _util
 from . import pytiledbsoma as clib
